+++ src/optimize.ts
@@ -8,7 +8,7 @@ import { ProgressPlugin, webpack } from 'webpack'
 import { merge } from 'webpack-merge'
 import { buildWConfig, envInit, toCtx, initCompilerLog } from './util'
 import { LANG, PLUGIN_NAME } from './const'
-import WebpackDevServer from 'webpack-dev-server'
+import WebpackDevServer, { Configuration as DevServerConfiguration } from 'webpack-dev-server'
 import { initMiddleWare } from 'yyl-base-webpack-config'
 
 const pkg = require('../package.json')
@@ -80,8 +80,24 @@ export const optimize: SeedOptimize = async (option: OptimizeOption) => {
         iRes.trigger('msg', ['info', [msg.toString()]])
       })
     }
-    iRes.trigger('progress', ['finished', 'success', [LANG.OPTIMIZE.LEGACY_POLYFILL_START]])
+    iRes.trigger('progress', ['finished', 'success', [LANG.OPTIMIZE.LEGACY_POLYFILL_FINISHED]])
+
+    // 不再支持 带 zepto 的支持 - 针对老旧项目
+    if (rootPkg.dependencies && rootPkg.dependencies.zepto) {
+      throw new Error(
+        `${LANG.OPTIMIZE.ZEPTO_NOT_SUPPORTED}: ${chalk.yellow('pkg.dependencies.zepto')}`
+      )
+    } else if (rootPkg.devDependencies && rootPkg.devDependencies.zepto) {
+      throw new Error(
+        `${LANG.OPTIMIZE.ZEPTO_NOT_SUPPORTED}: ${chalk.yellow('pkg.devDependencies.zepto')}`
+      )
+    } else if (yylConfig?.alias?.zepto) {
+      throw new Error(
+        `${LANG.OPTIMIZE.ZEPTO_NOT_SUPPORTED}: ${chalk.yellow('yylConfig.alias.zepto')}`
+      )
+    }
   }
+
   // - 运行前校验
 
   const wConfig = buildWConfig({
@@ -135,7 +151,7 @@ export const optimize: SeedOptimize = async (option: OptimizeOption) => {
           }
         })
       ]
-    })
+    } as any)
   )
 
   /** 使用项目自带server */
@@ -208,9 +224,12 @@ export const optimize: SeedOptimize = async (option: OptimizeOption) => {
           }
 
           try {
-            const devServer = new WebpackDevServer(compiler, {
-              ...wConfig.devServer
-            } as any)
+            const devServer = new WebpackDevServer(
+              compiler as any,
+              {
+                ...(wConfig.devServer as DevServerConfiguration)
+              } as any
+            )
             devServer.listen(serverPort, (err) => {
               if (err) {
                 iRes.trigger('msg', ['error', [LANG.OPTIMIZE.DEV_SERVER_START_FAIL, err]])